server/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "parana-server"
version = "0.1.0"
description = "REST API and chat service for the Parana coverage-tracking system"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.29",
    "psycopg[binary,pool]>=3.1",
    "python-dotenv>=1.0",
    "openai>=1.30",
]

[project.scripts]
parana-server = "parana_server.main:run"

[tool.hatch.build.targets.wheel]
packages = ["src/parana_server"]

[tool.hatch.build.targets.wheel.sources]
"src" = ""

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "httpx>=0.27",
    "testcontainers[postgres]>=4.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
